requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.3.1
requests==2.31.0
python-decouple==3.8